+++ RazorSvelte/App/chartdemo.svelte
@@ -1,0 +1,19 @@
+<script lang="ts">
+    import Layout from "./shared/layout/default";
+</script>
+
+<Layout>
+    <div class="container text-center">
+        <h1>ChartDemo</h1>
+        
+        <p>
+            Content here
+        </p>
+    </div>
+</Layout>
+
+<style lang="scss">
+    .container {
+        margin-top: 10%;
+    }
+</style>