+++ RazorSvelte/App/shared/urls.ts
@@ -4,15 +4,17 @@ export default {
     signInGitHubUrl: "/signin-github",
     signInGoogleUrl: "/signin-google",
     signInLinkedInUrl: "/signin-linkedin",
+    cardDemoUrl: "/carddemo",
+    chartDemoUrl: "/chartdemo",
     errorUrl: "/error",
     indexUrl: "/",
     loginUrl: "/login",
     logoutUrl: "/logout",
     notFoundUrl: "/404",
-    privacyUrl: "/privacy",
     spaUrl: "/spa",
     unathorizedUrl: "/401",
 
+    // endpoints
     chart1Url: "/api/chart/1",
     chart2Url: "/api/chart/2",
     chart3Url: "/api/chart/3",