+++ RazorSvelte/Pages/CardDemo.cshtml.cs
@@ -1,0 +1,7 @@
+namespace RazorSvelte.Pages;
+
+public partial class Urls
+{
+    public const string  CardDemoUrl = "/carddemo";
+}
+public class CardDemoModel : PageModel {}