+++ RazorSvelte/Pages/ChartDemo.cshtml.cs
@@ -1,0 +1,7 @@
+namespace RazorSvelte.Pages;
+
+public partial class Urls
+{
+    public const string  ChartDemoUrl = "/chartdemo";
+}
+public class ChartDemoModel : PageModel {}