+++ RazorSvelte/Pages/Privacy.cshtml.cs
@@ -1,9 +1,0 @@
-namespace RazorSvelte.Pages;
-
-public partial class Urls
-{
-    public const string PrivacyUrl = "/privacy";
-}
-
-public class PrivacyModel : PageModel { }
-